fix(codex): confirm early cancellation before cleanup
