+++ extensions/codex/src/app-server/attempt-client-cleanup.test.ts
@@ -1,4 +1,5 @@
 // Codex tests cover attempt client cleanup plugin behavior.
+import { setImmediate } from "node:timers/promises";
 import { describe, expect, it, vi } from "vitest";
 import {
   closeCodexStartupClientBestEffort,
@@ -8,6 +9,7 @@ import {
   terminateCodexBackgroundTerminals,
 } from "./attempt-client-cleanup.js";
 import { createClientHarness } from "./test-support.js";
+import { getCodexAppServerTurnRouter } from "./turn-router.js";
 
 describe("Codex app-server attempt client cleanup", () => {
   it.each([true, false])(
@@ -146,6 +148,256 @@ describe("Codex app-server attempt client cleanup", () => {
     );
   });
 
+  it.each(["", "turn-1"])(
+    "does not interrupt through an already-released route (%s)",
+    async (turnId) => {
+      const harness = createClientHarness();
+      try {
+        const route = getCodexAppServerTurnRouter(harness.client).reserveThread({
+          threadId: "thread-1",
+        });
+        route.release();
+        await expect(
+          interruptCodexTurnAndWaitBestEffort(harness.client, {
+            threadId: "thread-1",
+            turnId,
+            ownershipSignal: route.signal,
+          }),
+        ).resolves.toBe(false);
+        expect(harness.writes).toHaveLength(0);
+        expect(harness.client.getCloseError()).toBeUndefined();
+      } finally {
+        harness.client.close();
+      }
+    },
+  );
+
+  it.each([false, true])(
+    "cancels an acknowledged turn when activation races an early interrupt (%s)",
+    async (startBeforeError) => {
+      const harness = createClientHarness();
+      try {
+        const completion = interruptCodexTurnAndWaitBestEffort(harness.client, {
+          threadId: "thread-1",
+          turnId: "turn-1",
+          timeoutMs: 1_000,
+        });
+        const settled = vi.fn();
+        void completion.then(settled);
+        const first = JSON.parse(harness.writes.at(-1) ?? "{}") as { id: number };
+        const start = {
+          method: "turn/started",
+          params: { threadId: "thread-1", turn: { id: "turn-1", status: "inProgress" } },
+        };
+        if (startBeforeError) {
+          harness.send(start);
+        }
+        harness.send({
+          id: first.id,
+          error: { code: -32_600, message: "no active turn to interrupt" },
+        });
+        await setImmediate();
+        expect(settled).not.toHaveBeenCalled();
+
+        harness.send({
+          method: "turn/started",
+          params: { threadId: "peer-thread", turn: { id: "peer-turn", status: "inProgress" } },
+        });
+        await setImmediate();
+        expect(harness.writes).toHaveLength(startBeforeError ? 2 : 1);
+        if (!startBeforeError) {
+          harness.send(start);
+        }
+        await vi.waitFor(() => expect(harness.writes).toHaveLength(2));
+        const second = JSON.parse(harness.writes.at(-1) ?? "{}") as { id: number };
+        expect(second).toMatchObject({
+          method: "turn/interrupt",
+          params: { threadId: "thread-1", turnId: "turn-1" },
+        });
+        harness.send({ id: second.id, result: {} });
+        await setImmediate();
+        expect(settled).not.toHaveBeenCalled();
+        harness.send({
+          method: "turn/completed",
+          params: { threadId: "thread-1", turn: { id: "turn-1", status: "interrupted" } },
+        });
+        await expect(completion).resolves.toBe(true);
+        expect(harness.client.getCloseError()).toBeUndefined();
+      } finally {
+        harness.client.close();
+      }
+    },
+  );
+
+  it("keeps one deadline through activation and terminal confirmation", async () => {
+    vi.useFakeTimers();
+    const harness = createClientHarness();
+    const request = vi.spyOn(harness.client, "request");
+    try {
+      const completion = interruptCodexTurnAndWaitBestEffort(harness.client, {
+        threadId: "thread-1",
+        turnId: "turn-1",
+        timeoutMs: 100,
+      });
+      const settled = vi.fn();
+      void completion.then(settled);
+      const first = JSON.parse(harness.writes.at(-1) ?? "{}") as { id: number };
+      harness.send({
+        id: first.id,
+        error: { code: -32_600, message: "no active turn to interrupt" },
+      });
+      await vi.advanceTimersByTimeAsync(60);
+      harness.send({
+        method: "turn/started",
+        params: { threadId: "thread-1", turn: { id: "turn-1", status: "inProgress" } },
+      });
+      await vi.advanceTimersByTimeAsync(0);
+      expect(harness.writes).toHaveLength(2);
+      const rpcRejected = vi.fn();
+      void Promise.resolve(request.mock.results.at(-1)?.value).catch(rpcRejected);
+      await vi.advanceTimersByTimeAsync(39);
+      expect(settled).not.toHaveBeenCalled();
+      expect(rpcRejected).not.toHaveBeenCalled();
+      await vi.advanceTimersByTimeAsync(1);
+      expect(settled).toHaveBeenCalledExactlyOnceWith(false);
+      expect(rpcRejected).toHaveBeenCalledOnce();
+      await expect(completion).resolves.toBe(false);
+      expect(harness.client.getCloseError()).toBeUndefined();
+    } finally {
+      harness.client.close();
+      vi.useRealTimers();
+    }
+  });
+
+  it.each(["no active turn to interrupt", "expected active turn id turn-1 but found turn-2"])(
+    "preserves a terminal receipt that precedes an interrupt rejection: %s",
+    async (message) => {
+      const harness = createClientHarness();
+      try {
+        const completion = interruptCodexTurnAndWaitBestEffort(harness.client, {
+          threadId: "thread-1",
+          turnId: "turn-1",
+          timeoutMs: 100,
+        });
+        const first = JSON.parse(harness.writes.at(-1) ?? "{}") as { id: number };
+        harness.send({
+          method: "turn/started",
+          params: { threadId: "thread-1", turn: { id: "turn-1", status: "inProgress" } },
+        });
+        harness.send({
+          method: "turn/completed",
+          params: { threadId: "thread-1", turn: { id: "turn-1", status: "completed" } },
+        });
+        harness.send({
+          method: "turn/started",
+          params: { threadId: "thread-1", turn: { id: "turn-2", status: "inProgress" } },
+        });
+        harness.send({ id: first.id, error: { code: -32_600, message } });
+        await expect(completion).resolves.toBe(true);
+        expect(harness.writes).toHaveLength(1);
+        expect(harness.client.getCloseError()).toBeUndefined();
+      } finally {
+        harness.client.close();
+      }
+    },
+  );
+
+  it("does not interrupt after its deadline when timeout callbacks are delayed", async () => {
+    vi.useFakeTimers();
+    const harness = createClientHarness();
+    try {
+      const completion = interruptCodexTurnAndWaitBestEffort(harness.client, {
+        threadId: "thread-1",
+        turnId: "turn-1",
+        timeoutMs: 100,
+      });
+      const first = JSON.parse(harness.writes.at(-1) ?? "{}") as { id: number };
+      harness.send({
+        method: "turn/started",
+        params: { threadId: "thread-1", turn: { id: "turn-1", status: "inProgress" } },
+      });
+      vi.setSystemTime(Date.now() + 100);
+      harness.send({
+        id: first.id,
+        error: { code: -32_600, message: "no active turn to interrupt" },
+      });
+      await vi.advanceTimersByTimeAsync(0);
+      expect(harness.writes).toHaveLength(1);
+      await vi.advanceTimersByTimeAsync(100);
+      await expect(completion).resolves.toBe(false);
+      expect(harness.writes).toHaveLength(1);
+    } finally {
+      harness.client.close();
+      vi.useRealTimers();
+    }
+  });
+
+  it.each(["released", "closed"])(
+    "does not act on queued activation after its owner is %s",
+    async (ended) => {
+      const harness = createClientHarness();
+      try {
+        const route = getCodexAppServerTurnRouter(harness.client).reserveThread({
+          threadId: "thread-1",
+          onNotification: vi.fn(),
+        });
+        const completion = interruptCodexTurnAndWaitBestEffort(harness.client, {
+          threadId: "thread-1",
+          turnId: "turn-1",
+          timeoutMs: 100,
+          ownershipSignal: route.signal,
+        });
+        const first = JSON.parse(harness.writes.at(-1) ?? "{}") as { id: number };
+        harness.send({
+          method: "turn/started",
+          params: { threadId: "thread-1", turn: { id: "turn-1", status: "inProgress" } },
+        });
+        if (ended === "released") {
+          route.release();
+        } else {
+          harness.client.close();
+        }
+        harness.send({
+          id: first.id,
+          error: { code: -32_600, message: "no active turn to interrupt" },
+        });
+        await expect(completion).resolves.toBe(false);
+        expect(harness.writes).toHaveLength(1);
+        if (ended === "released") {
+          expect(harness.client.getCloseError()).toBeUndefined();
+        }
+      } finally {
+        harness.client.close();
+      }
+    },
+  );
+
+  it("uses retained bound-turn completion after a later cleanup request", async () => {
+    const harness = createClientHarness();
+    try {
+      const route = getCodexAppServerTurnRouter(harness.client).reserveThread({
+        threadId: "thread-1",
+        onNotification: vi.fn(),
+      });
+      route.armTurn();
+      await route.bindTurn("turn-1");
+      harness.send({
+        method: "turn/completed",
+        params: { threadId: "thread-1", turn: { id: "turn-1", status: "completed" } },
+      });
+      const completion = interruptCodexTurnAndWaitBestEffort(harness.client, {
+        threadId: "thread-1",
+        turnId: "turn-1",
+        timeoutMs: 100,
+      });
+      await expect(completion).resolves.toBe(true);
+      expect(harness.writes).toHaveLength(0);
+      route.release();
+    } finally {
+      harness.client.close();
+    }
+  });
+
   it("fails closed when acknowledged interruption never reaches its terminal", async () => {
     const harness = createClientHarness();
     const completion = interruptCodexTurnAndWaitBestEffort(harness.client, {
@@ -162,9 +414,14 @@ describe("Codex app-server attempt client cleanup", () => {
 
   it.each([
     {
-      name: "the exact already-terminal error",
+      name: "the absent-active error with a terminal receipt",
       error: { code: -32_600, message: "no active turn to interrupt" },
       completed: true,
+    },
+    {
+      name: "the absent-active error without a terminal receipt",
+      error: { code: -32_600, message: "no active turn to interrupt" },
+      completed: false,
     },
     {
       name: "another invalid-request error",
@@ -176,15 +433,21 @@ describe("Codex app-server attempt client cleanup", () => {
       error: { code: -32_000, message: "no active turn to interrupt" },
       completed: false,
     },
-  ])("classifies $name without waiting for a dropped terminal", async ({ error, completed }) => {
+  ])("confirms $name only from native evidence", async ({ error, completed }) => {
     const harness = createClientHarness();
     const completion = interruptCodexTurnAndWaitBestEffort(harness.client, {
       threadId: "thread-1",
       turnId: "turn-1",
-      timeoutMs: 1_000,
+      timeoutMs: 10,
     });
     const request = JSON.parse(harness.writes.at(-1) ?? "{}") as { id: number };
     harness.send({ id: request.id, error });
+    if (completed) {
+      harness.send({
+        method: "turn/completed",
+        params: { threadId: "thread-1", turn: { id: "turn-1", status: "completed" } },
+      });
+    }
 
     await expect(completion).resolves.toBe(completed);
     harness.client.close();
