+++ extensions/codex/src/app-server/attempt-client-cleanup.ts
@@ -2,6 +2,7 @@
  * Best-effort cleanup helpers for Codex app-server startup attempts and turns.
  */
 import { embeddedAgentLog } from "openclaw/plugin-sdk/agent-harness-runtime";
+import { createDeferred } from "openclaw/plugin-sdk/extension-shared";
 import { unsubscribeCodexAppServerLiveThread } from "./client-runtime.js";
 import { CodexAppServerRpcError, type CodexAppServerClient } from "./client.js";
 import { retireSharedCodexAppServerClientIfCurrent } from "./shared-client.js";
@@ -14,10 +15,8 @@ export const CODEX_APP_SERVER_UNSUBSCRIBE_TIMEOUT_MS = 5_000;
 const CODEX_NO_ACTIVE_TURN_ERROR_CODE = -32_600;
 const CODEX_NO_ACTIVE_TURN_ERROR_MESSAGE = "no active turn to interrupt";
 
-/** Identifies Codex's exact proof that an interrupt target already finished. */
-export function isCodexAlreadyTerminalInterruptError(
-  error: unknown,
-): error is CodexAppServerRpcError {
+/** Codex also reports this before an accepted turn publishes its start event. */
+export function isCodexNoActiveTurnInterruptError(error: unknown): error is CodexAppServerRpcError {
   return (
     error instanceof CodexAppServerRpcError &&
     error.code === CODEX_NO_ACTIVE_TURN_ERROR_CODE &&
@@ -95,6 +94,8 @@ export async function interruptCodexTurnAndWaitBestEffort(
     threadId: string;
     turnId: string;
     timeoutMs?: number;
+    /** Route-generation lifetime, distinct from the run abort being cleaned up. */
+    ownershipSignal?: AbortSignal;
   },
 ): Promise<boolean> {
   const timeoutMs =
@@ -102,27 +103,67 @@ export async function interruptCodexTurnAndWaitBestEffort(
       ? params.timeoutMs
       : CODEX_APP_SERVER_INTERRUPT_TIMEOUT_MS;
   const requestParams = { threadId: params.threadId, turnId: params.turnId };
-  let completion: { completion: Promise<boolean>; cancel: () => void } | undefined;
+  let cancelWatch: (() => void) | undefined;
   try {
-    // Codex acknowledges interruption before publishing turn/completed. Register
-    // first so an immediate exact-turn terminal cannot race past its owner.
-    completion = params.turnId
-      ? getCodexAppServerTurnRouter(client).watchNativeTurnCompletion({
-          threadId: params.threadId,
-          turnId: params.turnId,
-          timeoutMs,
-        })
-      : undefined;
-    await client.request("turn/interrupt", requestParams, { timeoutMs });
-    return completion ? await completion.completion : true;
-  } catch (error) {
-    if (isCodexAlreadyTerminalInterruptError(error)) {
+    params.ownershipSignal?.throwIfAborted();
+    if (!params.turnId) {
+      await client.request("turn/interrupt", requestParams, {
+        timeoutMs,
+        ...(params.ownershipSignal ? { signal: params.ownershipSignal } : {}),
+      });
       return true;
     }
+    const deadline = Date.now() + timeoutMs;
+    const started = createDeferred<boolean>();
+    // Codex acknowledges interruption before publishing turn/completed. Register
+    // first so an immediate exact-turn terminal cannot race past its owner.
+    const completion = getCodexAppServerTurnRouter(client).watchNativeTurnCompletion({
+      threadId: params.threadId,
+      turnId: params.turnId,
+      timeoutMs,
+      signal: params.ownershipSignal,
+      onStarted: () => started.resolve(true),
+    });
+    cancelWatch = completion.cancel;
+    if (completion.state !== "pending") {
+      return await completion.completion;
+    }
+    const requestInterrupt = async () => {
+      try {
+        await client.request("turn/interrupt", requestParams, {
+          timeoutMs: Math.max(1, deadline - Date.now()),
+          // The client floors RPC timeouts at 100ms. The lifecycle signal owns
+          // the exact remaining deadline and cancels RPCs when terminal wins.
+          signal: completion.settledSignal,
+        });
+        return true;
+      } catch (error) {
+        if (completion.state === "confirmed") {
+          return true;
+        }
+        if (isCodexNoActiveTurnInterruptError(error)) {
+          return false;
+        }
+        throw error;
+      }
+    };
+    if (!(await requestInterrupt())) {
+      // turn/start may acknowledge before native activation. Only that exact
+      // start receipt permits another interrupt; absent-active is not terminal proof.
+      const activated = await Promise.race([
+        completion.completion.then(() => false),
+        started.promise,
+      ]);
+      if (activated && completion.state === "pending" && Date.now() < deadline) {
+        await requestInterrupt();
+      }
+    }
+    return await completion.completion;
+  } catch (error) {
     embeddedAgentLog.debug("codex app-server turn interrupt failed during abort", { error });
     return false;
   } finally {
-    completion?.cancel();
+    cancelWatch?.();
   }
 }
